Make body have colour scheme's background colour
Make the body have the main background colour of the appropriate
colour scheme. This makes a page with a short bit of code (like a
one-line patch for instance) look a lot nicer, especially when using
a dark theme, as it reduces the contrast between the page and the code.