Parse the filename correctly
Remove the URL query parameters before extracting the
filename from the URL.

If the URL query parameters contain slashes, when the
function split the URL string, it got the filename wrongly.
Therefore, the syntax highligthing does not work properly.

For example:

https://gerrit.example.com/path/to/file.c?h=branch/name

was returning "name" as filename, when the expected
output is "file.c".

Signed-off-by: Javier Albornoz <javier.albornoz@hpe.com>